MM_33223 : Including a wrapper to DataOperator
