+++ app/database/admin/exceptions/database_connection_exception.ts
@@ -5,8 +5,8 @@
  * DatabaseConnectionException: This error can be thrown whenever an issue arises with the Database
  */
 class DatabaseConnectionException extends Error {
-    tableName: string;
-    constructor(message: string, tableName: string) {
+    tableName?: string;
+    constructor(message: string, tableName?: string) {
         super(message);
         this.tableName = tableName;
         this.name = 'DatabaseConnectionException';
