+++ app/database/admin/exceptions/database_operator_exception.ts
